requirements.txt
Pillow
emoji
requests
python-telegram-bot
requests-toolbelt
git+https://github.com/polarbearjngl/chatbase-python